This is kind of a large commit that adds a number of new functions.
BUG FIXES
---------
The hibernate-gtldb2.cfg.xml and hibernate-lut.cfg.xml files were pointing to the opposite servers.

NEW FEATURES
------------
Modified the server config files to talk to KBase instances of GLAMMdb (and added a hibernate-kbase.cfg.xml).

1. Metabolic Models
-------------------
The MetabolicModel class is used to represent the kinds of data that goes into metabolic models (e.g. FBA-style models), primarily Reactions and Compounds. This is similar to a MetabolicNetwork, but contains additional metadata for Reactions and Compounds. This includes compartment information for Compounds, and flux boundary information for Reactions.

This class is also represented in a MetabolicModelPresenter and MetabolicModelView. The View currently functions as just a little ListBox (above the AMDView). When a model is chosen in the ListBox, the presenter fetches its information through an rpc call that talks to GetMetabolicModel, MetabolicModelDAO, and MetabolicModelDAOImpl.

The other main change here is that when a MetabolicModel is loaded, it automatically loads any Reactions that are not already present in the main panel into the SVG as a kind of sidebar. These reactions behave (or, at least, should) as other reactions in the map, but are only present when the MetabolicModel is loaded. This necessitated changes to AnnotatedMapPresenter and AnnotatedMapData

There are also new prototype classes associated with things here: FluxExperiment and GetFluxes. A FluxExperiment just describes the metadata associated with an FBA experiment, and GetFluxes returns a set of Samples for the Reactions in a MetabolicNetwork. These should probably be just rolled into the Experiment/Sample data model somehow.


2. Group Data
-------------
The idea here is that some sets of reactions and/or compounds should be represented as belonging to one or more groups. For example, RegPrecise provides regulon data - a regulatory gene that has an effect on several other genes. Another example might be to look at sets of genes that are shared (or differ) between species.

GLAMM should be able to handle this data in the following way:
A. Input data either from a user-uploaded file or an external web service.
B. Parse data into groups of HasType objects with associated metadata (name, callback URL, group source)
C. Build a presenter that shows a list of all available groups and their metadata.
D. Paint the group information on the main map - each group of reactions/compounds should be a different color.
E. Incorporate interface widgets to toggle between which groups are visible.
F. Show which compounds/reactions belong to each group when clicked on (i.e. in their respective popup panel)

The code checked in to address this need is pretty early stage, and will likely get refactored into something that makes better sense later. For now, here's what's been done.

Client-side:
------------
OverlayDataGroup - the data model that holds the group data. This essentially holds a name, callback URL for the group, and set of HasTypes that represent the group members.
GroupDataPresenter/View - the main UI widget and controller for looking at the GroupData summaries in a table and loading from either a file or service
GroupDataUploadPresenter/View - widget and controller for uploading a data file (stub code for now)
GroupDataServicePresenter/View - widget and controller for downloading from an external data service.

RPC:
----
Added functions to GlammService and GlammServiceAsync to fetch OverlayGroupData

Server-side:
------------
GetGroupData, GroupDataDAO, GroupDataDAOImpl - accessory classes for fetching group data from a source.
GroupDataServiceManager - loads and populates a set of GroupDataServices from the external_services.xml file.
GroupDataService - contains information about the URL of the data service and the parameters involved.
PopulateDataServices - gets a list of what services have been loaded by the GroupDataServiceManager.

The design here is like this:
User inputs information into a the GroupDataServiceView.
GroupDataServicePresenter parses that and sends it to the server through an RPC.
The servlet uses a GroupDataService to format the parameters into a URL, and calls it.
The servlet then parses the response into a Set of OverlayGroupData (or an error) and returns this to the client side.
The client side interprets the response and shows it to the user.

In this commit, there's a little bit implemented.
A user can invoke a service and get GroupData (currently only from RegPrecise).
This GroupData is used to paint reactions and compounds in the main map.
It also is used to populate a table containing a legend of what is shown, with callback URLs to the source of each data group.

A future commit will implement toggling between which groups are being painted on screen, selecting a set of groups to show, and some way to show reactions that overlap between multiple groups (maybe some kind of alpha value in their color? Or some kind of css trickery that can show multiple colors via dashes?)

Future commits will also be much smaller.


NOTE:
-----
The Views for the Metabolic Model and Group Data work are currently hidden in this commit. The bug fixes in this (and the previous) commit will be made live soon, and the new functions are not ready.