1. Fixed bug where MetabolicModel SVG data wasn't being removed correctly. Also, it should now properly include the extra SVG elements when calculating total SVG size.
2. The server_config file was updated to include KBase DB information.

3. Added classes to define a program state. This allows interaction with the History mechanism, and defining URLs that put GLAMM in a specific state. E.g.: glamm.lbl.gov#taxId=511145 will initialize GLAMM with E. coli K-12 (taxonomy id 511145) as the species.

Currently, only, taxId=<MO species id> and mapId=<glamm map id.. map01100> are functional. Experiment, metabolic model, current viewport, will all come soon.

The process of digesting History tokens is implemented as an RPC call. The history token (looks like a URL anchor) is passed through an RPC. The server parses the token and constructs a GlammState object with necessary pieces - an Organism and String for the AMDescriptor. This is sent back to the client which sends the information around in the usual way.

To make this work on startup, the AppController.start() method has been split into start() and initialize(). In start(), some basic necessities are loaded up, then the history token is sent via RPC. On the RPC's success, initialize() is called with the resulting GlammState. This breaks the MVC paradigm a bit, as the Organism is passed around to the newly initialized presenters directly, instead of by Event. But as this is only done at the start, I think it's okay for now. When the history token is changed (user presses "back" or "forward" on the browser), GlammState components are passed around by firing events.