Hefty commit.
Added ability to view FBA models and results taken from KBase workspace services.
Refactored the login process to use the KBase authentication system.

An upcoming commit will include configuration to switch between deployment in KBase vs. deployment in LBL space.

This is also on the prototype-side and includes a number of classes that are copied from KBase. These will be culled in a future version.